Update README.md to reflect test/recipes structure
Recently, when starting a new project `kitchen-inspec` as my
verifier rather than `serverspec`, I noticed that it was looking in
`test/recipes/default` for tests`. This directory *did* exist because I
had created the cookbook with `chef generate cookbook`, but I had
written my initial test in `test/integration/default` out of habit.

In the commit history I noticed that d2b2742 changed the
`test_base_path` to `test/recipes` but it doesn't look like the
documentation was updated to reflect this.