Tweak logging for clarity
Do not print an info message saying that we are loading tests from the
test_base_dir unless we actually are. This can be confusing for
first-time users who may not even have that directory (if they used
ChefDK to generate the .kitchen.yml). Rather, iterate through the
targets that we _are_ testing and log those.

Signed-off-by: Tom Duffield <tom@chef.io>