Add support for pty config
Some hosts are configured to require a pty when using sudo. This will
result in the following error:

> Sudo failed: Sudo requires a TTY. Please see the README on how to
> configure sudo to allow for non-interactive usage.

It may not always be desirable to remove the `requiretty` flag in the
sudo configuration, so I am suggesting that we add the ability to use
a pty when connecting with SSH for those cases. Since this is already
supported in Train, all we need to do is expose that setting in
`.kitchen.yml`.

Signed-off-by: Christoffer Reijer <christoffer.brodd-reijer@basalt.se>