Require Ruby 2.3+ / InSpec 1.47+ and update dev deps (#221)
* Remove github_changelog_generator dep

We use expeditor now

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove bundler dep in the gemfile

This does nothing, but break things on bundler 2.x

Signed-off-by: Tim Smith <tsmith@chef.io>

* Require Ruby 2.3 or later

Remove support for EOL Ruby

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove unused Code Climate support

We're not using this anymore. Stop carrying around the baggage.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update rake and chefstyle

Bring our deps up to date.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Rename the kitchen file

No one likes hidden files.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Require InSpec 1.47+

Since we ship kitchen-inspec in DK for 99.9% of users they already have a fully up-to-date InSpec. This just allows us to remove a conditional and a test.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix capitalization of InSpec

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update ruby versions to test in Travis

Signed-off-by: Tim Smith <tsmith@chef.io>

* Force the bundler install in Travis and install the integration gems

We need these gems for testing in Travis

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update version to 1.0.0

Signed-off-by: Tim Smith <tsmith@chef.io>

* Properly limit the files we ship in the gem

The extra `|` was catching everything

Signed-off-by: Tim Smith <tsmith@chef.io>

* Test on xenial in Travis

Also disable sudo since they ignore this config now.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Ruby 2.6.0 testing to Travis

Make sure we test on everything we support

Signed-off-by: Tim Smith <tsmith@chef.io>